+++ bin/flutter_base_kit.dart
@@ -3,7 +3,7 @@
 import 'package:flutter_base_kit/generator/cli/arg_parser.dart';
 import 'package:flutter_base_kit/generator/cli/commands/create_command.dart';
 import 'package:flutter_base_kit/generator/cli/commands/init_command.dart';
-import 'package:universal_io/io.dart';
+import 'dart:io';
 
 void main(List<String> arguments) async {
   final parser = CliArgParser.createParser();